azure-medallion-pipeline/requirements.txt
azure-storage-file-datalake==12.14.0
azure-identity==1.15.0
psycopg2-binary==2.9.9
pandas==2.1.4
pyarrow==14.0.2
requests==2.31.0
great-expectations==0.18.8
pyspark==3.5.0
delta-spark==3.0.0
dbt-core==1.7.4
dbt-synapse==1.7.0
apache-airflow==2.8.0
apache-airflow-providers-microsoft-azure==8.3.0
python-dotenv==1.0.0
